Tested nesting directives containing SVG elements so we could recursively create orbiting objects around any other object and it could have gone as deeply as the model specified, but it seems to be impossible...
Angularjs create elements with the xhtml xmlns and not the svg one, so
chrome cannot parse the new svg elements. This will limit our capability
to create svg directives. The strange thing is with one directive
containing the svg it works. Things go bad when a svg directive is
inside another one.

I deleted the code i used for testing and just expanded what has been
done in the previous commit. I added mousewheel support for zoomin/out
and pressing mouse button + mousemove to move the star system around.

I need to add limits to the movements and i need to rework the zoom
functionalities to simplify the code.